feat: add new controller to register user
This controller uses the middleware that validates signup body.
It also uses the users repository to create a new user.
If the repository returns an error, it responds with 500 or 409 if the error is unique key violation.
Also changes the schema to use camel case instead of snake case, snake case is used only on repositories.