feat: add controller to login user
This controller gets the user by the email passed as credentials.
If the user does not exist or the password does not match the one on the database return 401.
Its important to keep the if statement in this order, because it will test the user existance first.
It also uses JWT to create a token based on the user id and return it.