feat: add controller to edit a post
This controller checks if the post exists and if the author is the same as the token passed before anything.
If it passes make the following:
edit the post;
clears unused urls from database;
delete old tag mentions from the previous post content;
create new mentions if there is any;
with the new relations, delete unused tags.